+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.11.5 (2021-04-15)
+* Updated Time and Number plugins to use Dictum interfaces
+* Simplified Time and Number plugin implementations
+
 ## v0.11.4 (2021-04-09)
 * Updated Systemic dependency